// src/redux/productsListPagination/productsListPagination.types.ts
export enum PRODUCTS_LIST_PAGINATION_ACTION_TYPES {
  SET_PRODUCTS = "productsListPagination/SET_PRODUCTS",
  SET_PAGE_NUMBER = "productsListPagination/SET_PAGE_NUMBER",
  SET_PAGE_SIZE = "productsListPagination/SET_PAGE_SIZE",
  FETCH_PRODUCTS_START = "productsListPagination/FETCH_PRODUCTS_START",
  FETCH_PRODUCTS_SUCCESS = "productsListPagination/FETCH_PRODUCTS_SUCCESS",
  FETCH_PRODUCTS_FAILED = "productsListPagination/FETCH_PRODUCTS_FAILED",
  FETCH_FILTERED_PRODUCTS_START = "productsListPagination/FETCH_FILTERED_PRODUCTS_START",
  FETCH_FILTERED_PRODUCTS_SUCCESS = "productsListPagination/FETCH_FILTERED_PRODUCTS_SUCCESS",
  FETCH_FILTERED_PRODUCTS_FAILED = "productsListPagination/FETCH_FILTERED_PRODUCTS_FAILED",
  INCREMENT_PAGE = "productsListPagination/INCREMENT_PAGE",
  DECREMENT_PAGE = "productsListPagination/DECREMENT_PAGE",
};